Wire up moved main-mobile.js
