+++ mobile/App.js
@@ -1,6 +1,6 @@
 import 'node-libs-react-native/globals';
 import Storybook from './storybook';
 
-import App from './main';
+import App from '../src/view/main-mobile';
 
 module.exports = !__DEV__ ? Storybook : App;
